Adjust vertical spacing between Enrollment and Metrics Data section cards
