+++ frontend/projects/upgrade/src/app/features/dashboard/experiments/pages/experiment-details-page/experiment-details-page-content/experiment-reward-feedback-section-card/experiment-reward-feedback-section-card.component.html
@@ -1,4 +1,4 @@
-<app-common-section-card *ngIf="shouldDisplayCard$ | async">
+<app-common-section-card>
   <app-common-section-card-title-header
     header-left
     [title]="'Reward Feedback'"
